requirements.txt
python-telegram-bot==21.10
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langchain-text-splitters==0.3.4
chromadb==0.6.3
openai==1.59.9
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
